+++ mono/utils/strenc.h
@@ -14,7 +14,7 @@
 #include <glib.h>
 #include <mono/utils/mono-publib.h>
 
-extern MONO_API gunichar2 *mono_unicode_from_external (const gchar *in, gsize *bytes);
+extern MONO_API gunichar2 *mono_unicode_from_external (const gchar *in, guint32 *bytes);
 extern MONO_API gchar *mono_utf8_from_external (const gchar *in);
 extern MONO_API gchar *mono_unicode_to_external (const gunichar2 *uni);
 extern MONO_API gboolean mono_utf8_validate_and_len (const gchar *source, glong* oLength, const gchar** oEnd);
